remorat assignments view
