+++ src/common/utils/iconDict.json
@@ -1,1 +1,1 @@
-["replits","arrowLeft2","screen","help","geekcoding","rocket","error","logo","arrowLeft","upload","homeFilled","longArrowRight","usaFlag","underlinedPencil","light","student","arrowDownFilled","conversation","dolarSign","success","renewal","sync","sideSupport","html","slack","arrowUp","globe","twitter","filter","calendar","linkedin","chronometer-full","sort","answer","running","gitpod","uploadImage","strength","regularFace","unchecked","pencil","idea","pdf","dolarSignBroke","youtube","google-collab","smile","image","copy","free","padlock","coding","close","people","learnpack","4Geeks-logo","checkboxChecked","facebook","camera","message","crown","cursorTooltip","setting","unverified","curvedLine","code","teacher","teacher1","codeBg","live-event","bell","flash","circle-check","socialMedia","live-event-opaque","checked2","book","info","add","infinite","card","warning","document","focus","logoModern","machine-learning","sad","certificate","key","instagram","4GeeksIcon","pencilFull","dot","closeRounded","arrowRight","git","download","dark","collab","graduationCap","group","close2","github","data-science-bg","logout","hamburger2","chronometer","switchLanguage","landing-avatars","assignments","arrowDown","onlinePeople","brain","file","community","verified","cameraFilled","leftArrow","data-science","spainFlag","ghost","bootstrap","eyeClosed","refresh_time","search","laptop-code","home","user","on-live","documentVerified","hamburger","eyeOpen","checked"]
+["search","dot","crown","copy","brain","replits","screen","slack","certificate","coding","cameraFilled","conversation","curvedLine","hamburger","unchecked","arrowDownFilled","pdf","sync","flash","image","light","html","dolarSignBroke","dark","document","eyeOpen","longArrowRight","eyeClosed","data-science-bg","student","upload","close2","arrowLeft","socialMedia","file","data-science","book","geekcoding","key","globe","4Geeks-logo","ghost","checkboxChecked","graduationCap","code","codeBg","help","bootstrap","leftArrow","focus","people","infinite","twitter","running","logo","arrowUp","home","regularFace","onlinePeople","landing-avatars","collab","dolarSign","4GeeksIcon","linkedin","laptop-code","instagram","sort","underlinedPencil","idea","checked2","user","cursorTooltip","free","circle-check","community","uploadImage","calendar","card","machine-learning","close","google-collab","chronometer-full","padlock","git","logout","bell","switchLanguage","closeRounded","checked","verified","youtube","pencil","rocket","warning","spainFlag","github","homeFilled","sideSupport","pencilFull","assignments","refresh_time","error","facebook","group","arrowLeft2","camera","success","learnpack","strength","smile","teacher","message","on-live","unverified","info","teacher1","gitpod","renewal","live-event","add","usaFlag","answer","sad","logoModern","documentVerified","hamburger2","arrowDown","setting","download","filter","arrowRight","live-event-opaque","chronometer"]
