+++ src/pages/cohort/[cohortSlug]/assignments.jsx
@@ -6,9 +6,6 @@ import useTranslation from 'next-translate/useTranslation';
 import {
   Box,
   Divider,
-  Avatar,
-  Flex,
-  useColorModeValue,
   useToast,
   Button,
   ButtonGroup,
@@ -28,26 +25,18 @@ import {
 } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
 import asPrivate from '../../../common/context/PrivateRouteWrapper';
-import ReactSelect from '../../../common/components/ReactSelect';
+import ReactSelect, { AsyncSelect } from '../../../common/components/ReactSelect';
 import Link from '../../../common/components/NextChakraLink';
 import Heading from '../../../common/components/Heading';
 import { usePersistent } from '../../../common/hooks/usePersistent';
 import bc from '../../../common/services/breathecode';
 import Icon from '../../../common/components/Icon';
 import Text from '../../../common/components/Text';
-import TaskLabel from '../../../common/components/taskLabel';
-import DottedTimeline from '../../../common/components/DottedTimeline';
 import useStyle from '../../../common/hooks/useStyle';
-import useFormatTimeString from '../../../common/hooks/useFormatTimeString';
-import { isGithubUrl } from '../../../utils/regex';
-import ButtonHandler, { ReviewModal, NoInfoModal, DeliverModal, DetailsModal } from '../../../js_modules/assignmentHandler/index';
 import useAssignments from '../../../common/store/actions/assignmentsAction';
-import { isWindow } from '../../../utils';
-// import Image from '../../../common/components/Image';
-import PopoverHandler from '../../../js_modules/assignmentHandler/PopoverHandler';
 import handlers from '../../../common/handlers';
-import LoaderScreen from '../../../common/components/LoaderScreen';
-import { ORIGIN_HOST } from '../../../utils/variables';
+import Projects from '../../../common/views/Projects';
+import StudentAssignments from '../../../common/views/StudentAssignments';
 
 function Assignments() {
   const { t } = useTranslation('assignments');
@@ -117,17 +106,13 @@ function Assignments() {
   const { query } = router;
   const { cohortSlug, academy } = query;
   const toast = useToast();
-  const { formatTimeString } = useFormatTimeString();
-  const { hexColor } = useStyle();
+  const { hexColor, borderColor2 } = useStyle();
   const { contextState, setContextState } = useAssignments();
   const [cohortSession] = usePersistent('cohortSession', {});
   const [syllabusData, setSyllabusData] = useState({
     assignments: [],
   });
   const [personalCohorts, setPersonalCohorts] = useState([]);
-  // const [allTasksPaginationProps, setAllTasksPaginationProps] = useState({});
-  const [allTasksOffset, setAllTasksOffset] = useState(20);
-  const [isFetching, setIsFetching] = useState(false);
   const [typeLabel, setTypeLabel] = useState(typeList.find((option) => option.value === query.task_type) || {
     label: t('type.project'),
     value: 'PROJECT',
@@ -142,8 +127,6 @@ function Assignments() {
   const [statusLabel, setStatusLabel] = useState(statusList.find((option) => option.value === query.status));
   const [openFilter, setOpenFilter] = useState(false);
   const [currentView, setCurrentView] = useState(Number(query.view) || 0);
-  const [currentTask, setCurrentTask] = useState(null);
-  const [deliveryUrl, setDeliveryUrl] = useState('');
   const [sort, setSort] = useState(query.sort || undefined);
 
   const [currentStudentList, setCurrentStudentList] = useState([]);
@@ -154,29 +137,6 @@ function Assignments() {
     status: 'loading',
   });
 
-  const linkColor = useColorModeValue('blue.default', 'blue.300');
-  const borderColor = useColorModeValue('gray.200', 'gray.500');
-
-  const lang = {
-    es: '/es/',
-    en: '/',
-  };
-
-  const statusColors = {
-    APPROVED: hexColor.green,
-    REJECTED: hexColor.danger,
-    UNDELIVERED: hexColor.danger,
-    DELIVERED: hexColor.yellowDefault,
-  };
-
-  const queryStudentExists = query.student !== undefined && query.student?.length > 0;
-  const queryStatusExists = query.status !== undefined && query.status?.length > 0;
-  const queryProjectExists = query.project !== undefined && query.project?.length > 0;
-
-  const studentDefaultValue = currentStudentList
-    ?.filter((l) => l.user.id === Number(router.query.student))
-    .map((l) => l?.user?.id)[0];
-
   const reverseStatus = (status) => {
     if (status === 'delivered') return { task_status: 'DONE', revision_status: 'PENDING' };
     if (status === 'approved') return { revision_status: 'APPROVED' };
@@ -186,12 +146,12 @@ function Assignments() {
     return {};
   };
 
-  const getFilterAssignments = (cohortId, academyId, studentId) => {
+  const getFilterAssignments = (cohortId, academyId) => {
     setLoadStatus({ loading: true, status: 'loading' });
     bc.todo({
       limit: 1000,
       task_type: typeLabel?.value || undefined,
-      student: studentId || undefined,
+      student: router.query.student || undefined,
       sort,
       educational_status: educationalLabel.length > 0 ? educationalLabel.map((val) => val.value).join(',').toUpperCase() : undefined,
       like: query.project,
@@ -199,7 +159,6 @@ function Assignments() {
     })
       .getAssignments({ id: cohortId, academy: academyId })
       .then((projectList) => {
-        setIsFetching(false);
         const allTasks = projectList.data?.results;
 
         setContextState({
@@ -207,7 +166,6 @@ function Assignments() {
         });
       })
       .catch((error) => {
-        setIsFetching(false);
         toast({
           position: 'top',
           title: t('alert-message:error-fetching-tasks'),
@@ -258,10 +216,9 @@ function Assignments() {
           if (syllabusInfo?.data) {
             let assignments = syllabusInfo.data.json.days.filter((obj) => obj.assignments && Array.isArray(obj.assignments) && obj.assignments.length > 0 && typeof obj.assignments[0] === 'object').map((obj) => obj.assignments);
             assignments = [].concat(...assignments);
-            const syllabusProjects = syllabusInfo.data.json.days.filter((day) => day.project && typeof day.project === 'object').map(({ project }) => ({ ...project }));
             if (query.project) {
-              const filteredProject = syllabusProjects.find((project) => project.slug === query.project);
-              setProjectLabel(filteredProject && { label: filteredProject.title, value: filteredProject.slug });
+              const filteredProject = assignments.find((project) => project.slug === query.project);
+              if (filteredProject) setProjectLabel({ label: filteredProject.title, value: filteredProject.slug });
             }
             setSyllabusData({
               assignments,
@@ -300,79 +257,36 @@ function Assignments() {
   }, [selectedCohort]);
 
   const loadStudents = () => {
-    if (selectedCohort) {
-      setLoadStatus({ loading: true, status: 'loading' });
-      const cohortId = selectedCohort.value;
-      const academyId = selectedCohort.academy || academy;
-      const slug = selectedCohort.slug || cohortSlug;
-      bc.cohort({ sort, users: query.student, educational_status: educationalLabel.length > 0 ? educationalLabel.map((val) => val.value).join(',').toUpperCase() : undefined })
-        .getStudentsWithTasks(slug, academyId)
-        .then((res) => {
-          const students = res?.data;
-          setCurrentStudentList(students);
-        })
-        .catch(() => {
-          toast({
-            position: 'top',
-            title: t('alert-message:error-fetching-students'),
-            status: 'error',
-            duration: 7000,
-            isClosable: true,
-          });
+    setLoadStatus({ loading: true, status: 'loading' });
+    const academyId = selectedCohort.academy || academy;
+    const { slug } = selectedCohort;
+    bc.cohort({ sort, users: query.student, educational_status: educationalLabel.length > 0 ? educationalLabel.map((val) => val.value).join(',').toUpperCase() : undefined })
+      .getStudentsWithTasks(slug, academyId)
+      .then((res) => {
+        const students = res?.data;
+        setCurrentStudentList(students);
+      })
+      .catch(() => {
+        toast({
+          position: 'top',
+          title: t('alert-message:error-fetching-students'),
+          status: 'error',
+          duration: 7000,
+          isClosable: true,
         });
-      getFilterAssignments(cohortId, academyId, router.query.student);
-    }
-    if (!selectedCohort?.value && selectedCohort !== null) {
-      setLoadStatus({ loading: false, status: 'idle' });
-    }
+      })
+      .finally(() => setLoadStatus({ loading: false, status: 'idle' }));
   };
 
   useEffect(() => {
-    loadStudents();
+    if (selectedCohort) {
+      loadStudents();
+      getFilterAssignments(selectedCohort.value, selectedCohort.academy || academy);
+    }
   }, [
     selectedCohort,
-    studentDefaultValue,
     router.query,
   ]);
-
-  const filteredTasks = contextState.allTasks.length > 0
-    ? contextState.allTasks
-      .filter((_, i) => i < allTasksOffset)
-    : [];
-
-  const handleScroll = () => {
-    const scrollTop = isWindow && document.documentElement.scrollTop;
-    const offsetHeight = isWindow && document.documentElement.offsetHeight + 15;
-    const innerHeight = isWindow && window.innerHeight;
-    if (innerHeight + scrollTop <= offsetHeight && !queryStudentExists) return;
-    setIsFetching(true);
-  };
-
-  useEffect(() => {
-    if (allTasksOffset < contextState.allTasks.length) {
-      window.addEventListener('scroll', handleScroll);
-      return () => window.removeEventListener('scroll', handleScroll);
-    }
-
-    return () => {};
-  }, [allTasksOffset, contextState]);
-
-  useEffect(() => {
-    if (!isFetching) return;
-    if (queryStudentExists || queryProjectExists || queryStatusExists) {
-      setAllTasksOffset(contextState.allTasks.length);
-      setIsFetching(false);
-    }
-
-    if (filteredTasks && allTasksOffset < contextState.allTasks.length) {
-      if (!queryStudentExists) {
-        setAllTasksOffset(allTasksOffset + 20);
-        setIsFetching(false);
-      }
-    } else {
-      setIsFetching(false);
-    }
-  }, [isFetching, queryStatusExists, queryProjectExists, queryStudentExists]);
 
   const closeFilterModal = () => setOpenFilter(false);
 
@@ -419,50 +333,6 @@ function Assignments() {
     loadStudents();
   };
 
-  const getStatus = (task) => {
-    if (!task) return null;
-    if (task.task_status === 'DONE' && task.revision_status === 'PENDING') return 'DELIVERED';
-    if (task.task_status === 'PENDING' && task.revision_status === 'PENDING') return 'UNDELIVERED';
-    return task.revision_status;
-  };
-
-  const showSingleTask = async (task) => {
-    try {
-      const status = getStatus(task);
-      let file;
-      const academyId = selectedCohort?.academy || academy || personalCohorts.find((l) => l.slug === cohortSlug)?.academy;
-      if (status === 'UNDELIVERED' || status === 'REJECTED') {
-        const { data } = await bc.todo().deliver({
-          id: task.id,
-          academy: academyId,
-        });
-        setDeliveryUrl(data.delivery_url);
-      } else if (status === 'APPROVED' || status === 'DELIVERED') {
-        setIsFetching(true);
-        const assetResp = await bc.lesson().getAsset(task.slug);
-        if (assetResp && assetResp.status < 400) {
-          const dataAsset = assetResp.data;
-          if (!dataAsset?.delivery_formats.includes('url')) {
-            const fileResp = await bc.todo().getFile({ id: task.id, academyId: cohortSession?.academy?.id });
-            if (fileResp && fileResp.status < 400) {
-              file = await fileResp.data;
-            }
-          }
-        }
-        setIsFetching(false);
-      }
-      setCurrentTask({ ...task, status, file });
-    } catch (e) {
-      toast({
-        position: 'top',
-        title: t('alert-message:review-url-error'),
-        status: 'error',
-        duration: 6000,
-        isClosable: true,
-      });
-    }
-  };
-
   return (
     <>
       <Box
@@ -476,7 +346,7 @@ function Assignments() {
         >
           <Link
             href={cohortSession?.selectedProgramSlug || '/choose-program'}
-            color={linkColor}
+            color={hexColor.blueDefault}
             display="inline-block"
             letterSpacing="0.05em"
             fontWeight="700"
@@ -532,7 +402,7 @@ function Assignments() {
           )}
         </Box>
       </Box>
-      <Divider borderBottom="1px solid" color={borderColor} />
+      <Divider borderBottom="1px solid" color={borderColor2} />
 
       <Box
         display="flex"
@@ -802,275 +672,15 @@ function Assignments() {
         p="0 0 30px 0"
       >
         {currentView === 1 ? (
-          <Box
-            minHeight="34vh"
-            borderRadius="3px"
-            margin="0 auto"
-            maxWidth="1012px"
-            flexGrow={1}
-            overflow="auto"
-          >
-            <Box
-              display="flex"
-              margin="20px 32px 20px 55px"
-              gridGap="10px"
-              justifyContent="space-between"
-              flexDirection="row"
-              alignItems="center"
-              width={{ base: 'max-content', md: 'auto' }}
-            >
-              <Text
-                size="15px"
-                display="flex"
-                width={{ base: '6.8rem', md: '50%' }}
-                fontWeight="700"
-              >
-                {t('label.status')}
-              </Text>
-              <Text
-                size="15px"
-                display="flex"
-                width={{ base: '13rem', md: '100%' }}
-                fontWeight="700"
-              >
-                {t('label.student-and-assignments')}
-              </Text>
-              <Text
-                size="15px"
-                display="flex"
-                width={{ base: '8.3rem', md: '40%' }}
-                fontWeight="700"
-              >
-                {t('label.link')}
-              </Text>
-              <Text
-                size="15px"
-                display="flex"
-                width={{ base: '25%', md: '25%' }}
-                minWidth="115px"
-                fontWeight="700"
-              >
-                {t('label.actions')}
-              </Text>
-            </Box>
-            <Box display="flex" flexDirection="column" gridGap="18px">
-              {filteredTasks.length > 0 ? (
-                filteredTasks.map((task, i) => {
-                  const index = i;
-                  const githubUrl = task?.github_url;
-                  const haveGithubDomain = githubUrl && isGithubUrl.test(githubUrl);
-                  const fullName = `${task.user.first_name} ${task.user.last_name}`;
-                  const projectLink = `${ORIGIN_HOST}${
-                    lang[router.locale]
-                  }project/${task.associated_slug}`;
-
-                  return (
-                    <Box
-                      key={`${index}-${task.slug}-${task.title}-${fullName}`}
-                      p="18px 28px"
-                      display="flex"
-                      width={{ base: 'max-content', md: '100%' }}
-                      minWidth={{ base: '620px', md: '100%' }}
-                      maxWidth={{ base: '620px', md: '100%' }}
-                      gridGap="10px"
-                      justifyContent="space-between"
-                      flexDirection="row"
-                      alignItems="center"
-                      border="1px solid"
-                      borderColor={borderColor}
-                      borderRadius="17px"
-                    >
-                      <Box
-                        display="flex"
-                        width="auto"
-                        minWidth="calc(160px - 0.5vw)"
-                        // width="153px"
-                      >
-                        <Box width="28px" height="28px" marginRight="15px">
-                          {syllabusData.assignments.find(
-                            (assignment) => assignment.slug === task.associated_slug && assignment.mandatory,
-                          ) && (
-                            <Icon
-                              icon="warning"
-                              color="yellow.default"
-                              width="28px"
-                              height="28px"
-                            />
-                          )}
-                        </Box>
-                        <TaskLabel currentTask={task} t={t} />
-                      </Box>
-
-                      <Box width="35%">
-                        <Text size="15px">{fullName}</Text>
-                        <Link
-                          variant="default"
-                          href={projectLink}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                        >
-                          {task.title}
-                        </Link>
-                      </Box>
-
-                      <PopoverHandler
-                        task={task}
-                        haveGithubDomain={haveGithubDomain}
-                        githubUrl={githubUrl}
-                      />
-
-                      <Box width="auto" minWidth="160px" textAlign="end">
-                        <ButtonHandler
-                          currentTask={task}
-                          cohortSession={cohortSession}
-                          contextState={contextState}
-                          updpateAssignment={updpateAssignment}
-                        />
-                      </Box>
-                    </Box>
-                  );
-                })
-              ) : (
-                <>
-                  {loadStatus.status === 'loading' && (
-                    <Box
-                      display="flex"
-                      justifyContent="center"
-                      mt="2rem"
-                      mb="5rem"
-                      position="relative"
-                    >
-                      <LoaderScreen width="80px" height="80px" />
-                    </Box>
-                  )}
-                  {loadStatus.loading === false && (
-                    <Text
-                      size="25px"
-                      pt="3rem"
-                      textAlign="center"
-                      display="flex"
-                      width="auto"
-                      margin="0 auto"
-                      fontWeight="700"
-                    >
-                      {t('common:search-not-found')}
-                    </Text>
-                  )}
-                </>
-              )}
-              {allTasksOffset < contextState.allTasks.length !== null
-                && isFetching && (
-                  <Box
-                    display="flex"
-                    justifyContent="center"
-                    mt="2rem"
-                    mb="5rem"
-                  >
-                    <LoaderScreen width="80px" height="80px" />
-                  </Box>
-              )}
-            </Box>
-          </Box>
+          <Projects updpateAssignment={updpateAssignment} loadStatus={loadStatus} syllabusData={syllabusData} />
         ) : (
-          <Box
-            minHeight="34vh"
-            borderRadius="3px"
-            margin="0 auto"
-            maxWidth="1012px"
-            flexGrow={1}
-            overflow="auto"
-          >
-            <Flex flexDirection="column" gridGap="18px">
-              {currentStudentList.map((student) => {
-                const { user } = student;
-                const fullname = `${student.user.first_name} ${student.user.last_name}`;
-                const percentage = Math.round((student.tasks.reduce((acum, val) => (val.task_status !== 'PENDING' && val.task_type === 'PROJECT' ? acum + 1 : acum), 0) / syllabusData.assignments.length) * 100);
-                const lastDeliver = student.tasks.reduce((date, val) => {
-                  if (date) return date > val ? date : val.updated_at;
-                  if (val.updated_at && val.task_status !== 'PENDING' && val.task_type === 'PROJECT') return val.updated_at;
-                  return null;
-                }, null);
-                const dots = syllabusData.assignments.map((elem) => {
-                  const studentTask = student.tasks.find((task) => task.associated_slug === elem.slug);
-                  const { mandatory } = elem;
-                  return {
-                    ...elem,
-                    ...studentTask,
-                    label: elem.title,
-                    highlight: mandatory,
-                    user,
-                    color: statusColors[getStatus(studentTask)] || 'gray',
-                  };
-                });
-                return (
-                  <DottedTimeline
-                    key={student.id}
-                    onClickDots={showSingleTask}
-                    label={(
-                      <Flex gridGap="10px" alignItems="center">
-                        <Avatar
-                          src={student.user.profile?.avatar_url}
-                          width="25px"
-                          height="25px"
-                          style={{ userSelect: 'none' }}
-                        />
-                        <Box>
-                          <p>{fullname}</p>
-                          <small>{`${percentage}${t('delivered-percentage')} - `}</small>
-                          <small>{lastDeliver ? t('last-deliver', { date: formatTimeString(new Date(lastDeliver)) }) : t('no-deliver')}</small>
-                        </Box>
-                      </Flex>
-                    )}
-                    dots={dots}
-                    helpText={`${t('educational-status')}: ${student.educational_status}`}
-                  />
-                );
-              })}
-            </Flex>
-            <ReviewModal
-              currentTask={currentTask}
-              projectLink={`${ORIGIN_HOST}${
-                lang[router.locale]
-              }project/${currentTask?.slug}`}
-              updpateAssignment={updpateAssignment}
-              isOpen={currentTask && currentTask.status === 'DELIVERED'}
-              onClose={() => setCurrentTask(null)}
-            />
-            <NoInfoModal
-              isOpen={currentTask && !currentTask.status}
-              onClose={() => setCurrentTask(null)}
-            />
-            <DeliverModal
-              currentTask={currentTask}
-              projectLink={`${ORIGIN_HOST}${
-                lang[router.locale]
-              }project/${currentTask?.slug}`}
-              updpateAssignment={updpateAssignment}
-              isOpen={currentTask && (currentTask.status === 'UNDELIVERED' || currentTask.status === 'REJECTED')}
-              onClose={() => setCurrentTask(null)}
-              deliveryUrl={deliveryUrl}
-            />
-            <DetailsModal
-              currentTask={currentTask}
-              projectLink={`${ORIGIN_HOST}${
-                lang[router.locale]
-              }project/${currentTask?.slug}`}
-              updpateAssignment={updpateAssignment}
-              isOpen={currentTask && currentTask.status === 'APPROVED'}
-              onClose={() => setCurrentTask(null)}
-            />
-            {loadStatus.status === 'loading' && (
-              <Box
-                display="flex"
-                justifyContent="center"
-                mt="2rem"
-                mb="5rem"
-                position="relative"
-              >
-                <LoaderScreen width="80px" height="80px" />
-              </Box>
-            )}
-          </Box>
+          <StudentAssignments
+            currentStudentList={currentStudentList}
+            loadStatus={loadStatus}
+            selectedCohort={selectedCohort}
+            syllabusData={syllabusData}
+            updpateAssignment={updpateAssignment}
+          />
         )}
       </Box>
     </>
